generate

+++ migrations/frontend/squashed.sql
@@ -4245,12 +4245,12 @@ CREATE TABLE repo_embedding_job_stats (
     code_files_embedded integer DEFAULT 0 NOT NULL,
     code_chunks_embedded integer DEFAULT 0 NOT NULL,
     code_files_skipped jsonb DEFAULT '{}'::jsonb NOT NULL,
-    code_bytes_embedded integer DEFAULT 0 NOT NULL,
+    code_bytes_embedded bigint DEFAULT 0 NOT NULL,
     text_files_total integer DEFAULT 0 NOT NULL,
     text_files_embedded integer DEFAULT 0 NOT NULL,
     text_chunks_embedded integer DEFAULT 0 NOT NULL,
     text_files_skipped jsonb DEFAULT '{}'::jsonb NOT NULL,
-    text_bytes_embedded integer DEFAULT 0 NOT NULL,
+    text_bytes_embedded bigint DEFAULT 0 NOT NULL,
     code_chunks_excluded integer DEFAULT 0 NOT NULL,
     text_chunks_excluded integer DEFAULT 0 NOT NULL
 );
